WA-4002 : Refactor to move all consent convenience methods into a separate class.
